+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
   - Update buildozer, p4a, and ndk versions, refs #21
   - Publish to F-Droid, refs #20
+  - Improve Android camera permission handling, refs #12, #16
 
 ## [20190902]